third_party/WebKit/LayoutTests/http/tests/webfont/webfont-performance-duration-expected.txt

CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
PASS successfullyParsed is true

TEST COMPLETE
PASS entry.startTime >= waitMillisBeforeFetching is true
Hello webfont
